Als 2548 508 compliance for links opening in new tab (#177)
* Add aria-label for target="_blank" links.

Including an aria-label for each target="_blank" allows screen readers to notify the user that the link will redirect them to a new tab.

* Add redirect confirmation modal

When a user clicks a link a modal is displayed with a message and two buttons. The buttons are "close" and "continue". If a user closes the modal nothing happens. If the user clicks continue they will continue to be redirected.

* Update link to settings.pdfLink