Update router getInvalidActiveFilters (#188)
* Update router getInvalidActiveFilters

Updating getInvalidActiveFilters we now check if they at least 'Gene_with_variant' in their query scope. We do not restrict genomic filtering based on studies so we cannot verify if they are permitted to access a specific gene like we can with selected variables.

* Update router getInvalidActiveFilters

Instead of using a list we are just checking for "Gene_with_variant".