Als 4472 leaving bdc dialog window closes the previous modal (#202)
* Change modal and fix handle tabs

* Update to handle different modal containers

By allowing a different modal ID to be passed we can specify with in which modal we intend to press the close button.

* Update to make modal container specific

The default modal is still the "modal-window" ensuring that current functionality does not break across the page. By allowing this we can have 1 or more modals displayed within different containers. It will allow the modal to work as if the other modal does not exist.

* Update footer.js

Moved redirectModal functionality to its own file. Updated the $(document) to first remove any on click events for a href that has target blank. Then applies a new event listener. This fixes a bug where occasionally multiple events are attached to 1 href.

* Update modal.js

backdrop is now static. This is done to handle a known bug where clicking off a modal closed ALL modals that were opened. Now we handle determining if someone clicked outside the modal and if they did we click the close button using jquery.

* Update to use new redirect-modal view

* Moved redirect modal to its own backbone view

* Remove console log

* Minor update to modal-backdrop

We now remove the modal-backdrop upon opening a new modal

* Fix redirect-modal probably handling backdrop

* Fix clicking outside the modal to close

* Update modal.js

When clicking on a button inside the modal that didn't have a direct parent of .modal-content. We now traverse the tree up until we find .modal-content or body. If we traverse all the way up to body we know we aren't in a modal.