Als 4615 (#211)
* Update redirect-modal.js

When the redirect modal was closed we appended a modal backdrop to the body if the source event parent was a modal. We needed to add an on event listener to ensure we also close the modal-backdrop on close.

* Fix studiesModal

Now correctly shows <10 and +- 3 values as part of the studies with counts. Was previously displayed in studies with 0 participants.