Render template earlier in landing.js (#265)
The order of operations has been altered in the landing.js file of the BioDataCatalyst UI. The current changes involve rendering the template right after the function is invoked rather than waiting until the end. This reordering will allow the loading icons to display until all ajax request are complete.