[ALS-6223] Refactor spinner management in visualization modal view
The 'spinning' attribute is no longer directly manipulated, enhancing code maintainability. A deferred object method now manage the spinner status. The previous approach was causing the application to continuously re-render the visualization section. Everytime the spinner was changed it would force a re-render of the view causing duplicate graphs to be displayed.