[ALS-6223] Duplicate graphs on BDC StatViz (#297)
* [ALS-6223] Modified plotlyjs loading to use a promise
Load plotly now returns a promise that is resolved upon plotly loading. This required moving all of the remaining logic into the "then" chain.

* [ALS-6223] Clear visualization container before rendering
* [ALS-6223] Empty visualization container after its all rendered
* [ALS-6223] Update getImages function to return Promise
Code changes are made to the getImages function to return a Promise. This allows for better sequencing of operations, and now images are loaded before the creation of visualizations. This addresses potential previously arising issues with images not loading in time for visualization creation.

* [ALS-6223] Refactor spinner management in visualization modal view
The 'spinning' attribute is no longer directly manipulated, enhancing code maintainability. A deferred object method now manage the spinner status. The previous approach was causing the application to continuously re-render the visualization section. Everytime the spinner was changed it would force a re-render of the view causing duplicate graphs to be displayed.