[ALS-6308] Add stigmatization status to variable info modal
Added a condition to check if a data variable is open access in the variable info modal. If it is, the user is given an option to report that the variable is not stigmatizing. If it isn't open access, a simple message is displayed stating that the variable is stigmatizing.