Release (#310)
* [ALS-5760] Add ternary operator to handle null case

* Minor H2 html element fix

* [ALS-5583] Update build process to support AIM-AHEAD PIC-SURE (#273)

* Add userDetails override
* Add psama

* [ALS-5584] AIM-AHEAD PIC-SURE: Hide the genomic filter
* [ALS-5567] Fix header links
* [ALS-5583] Optimized Dockerfile and added separate Dockerfile for UI overrides
Refactored the biodatacatalyst-ui Dockerfile to remove unnecessary build stages and environment variable (IS_OPEN_ACCESS). Also created a separate Dockerfile for future UI overrides. The changes aim to increase efficiency in the deployment process of UI components and make the Dockerfile more understandable.

---------

Co-authored-by: James <Jamestp19@gmail.com>

* [ALS-5703] AIM-AHEAD PIC-SURE Help Tab (#278)

* [ALS-5703] Update dropdown controls

No longer toggles the wrong dropdown menu if two menus are present in the header. Now calculates if the dropdown menu would go beyond the viewport's width. If it does, it recalculates the left property to align the dropdown to the right edge of the screen.

* [ALS-5703] dropdown is now focused correctly

* [ALS-5789] Add Loading spinner when search result selected (#279)

* Improve redirectModal message flexibility

In the redirect-modal.js file, the headerMessage and message have been refactored to accept dynamic inputs from the calling method. Similarly, the footer.js has been modified to pass a custom message to the redirectModal. This enhancement increases flexibility and allows customized messages for different external links.

* [ALS-5883] Allow optional override of download modal message

* [ALS-5584] Add ontology instance dependency to search-view (#281)

* [ALS-5584] Add ontology instance dependency to search-view 
This refactoring rearranges code order in search-view.js and modifies the render function to enable genomic filtering based on ontology's allInfoColumns condition. This allows the display of genomic filtering options to be dependent on the state of genomic data.

* [ALS-5332] Add Admin consoles and Fix bug with saving a user (#282)

* Add psama

* Add userDetails override

* Fix missing imports and add better checks

* [ALS-5077] Remove "Counts" column from studies table (#283)

* [ALS-5082] Persist database: Flyway (#287)

* [ALS-5082] Replace hardcoded resource id with configurable setting
* [ALS-6058] Refactor search term selection in search view (#286)
* [ALS-5077] Remove "Counts" column from studies table (#283)

* [ALS-5548] Stigmatizing Variables Displayed in Dataset View of Open PIC-SURE (#285)

* Filter out stigmatized variables in datatable

The update ensures that any stigmatized variable isn't displayed in the datatable. This is done by filtering the model by checking if the 'is_stigmatized' property of each variable is set to false before mapping the rest of the properties. Stigmatization status is determined by the variable's metadata.

* [ALS-5548] Filter for Open Access only

* [ALS-6144] Fix filter variable info modal button (#289)

* Update condition for variable display in datatable-filter-modal-view (#291)

* Update condition for variable display in datatable-filter-modal-view

* Refactor condition logic in datatable-filter-modal-view.js for variable display when access is open. Variables that are tagged as 'stigmatized' will be filtered out when access is open; otherwise, all variables will be displayed. This modifies the system behaviour by adding a conditional check for the 'isOpenAccess' session variable.

* "Conditionally add BioData Catalyst-specific filters" (#290)

Added conditions in filter-model.js to only include the Parent Study Accession and TOPMed Study Accession with Subject ID filters if 'include_bdc_specific_filters' is true in settings. Enhanced control over filtering study-specific data and improved system flexibility by allowing toggling of this functionality.

* Merge for AIM-AHEAD 2024-03-14-Release (#288) (#293)

[ALS-5082] Persist database: Flyway

* [ALS-6186] Fix filter text length (#294)

* [ALS-6165] Disable directory browsing  (#296)

* "Conditionally add BioData Catalyst-specific filters"
Added conditions in filter-model.js to only include the Parent Study Accession and TOPMed Study Accession with Subject ID filters if 'include_bdc_specific_filters' is true in settings. Enhanced control over filtering study-specific data and improved system flexibility by allowing toggling of this functionality.

* Disable directory browsing in Dockerfile
The Dockerfile has been updated to disable directory browsing in the Apache HTTPD server configuration. This improves the server's security by preventing unwanted access to directory listings. It was done by replacing the 'Options Indexes FollowSymLinks' line in the httpd.conf file with 'Options -Indexes FollowSymLinks'.

* Adjust Apache configuration in Dockerfiles

* [ALS-6223] Duplicate graphs on BDC StatViz (#297)

* [ALS-6223] Modified plotlyjs loading to use a promise
Load plotly now returns a promise that is resolved upon plotly loading. This required moving all of the remaining logic into the "then" chain.

* [ALS-6223] Clear visualization container before rendering
* [ALS-6223] Empty visualization container after its all rendered
* [ALS-6223] Update getImages function to return Promise
Code changes are made to the getImages function to return a Promise. This allows for better sequencing of operations, and now images are loaded before the creation of visualizations. This addresses potential previously arising issues with images not loading in time for visualization creation.

* [ALS-6223] Refactor spinner management in visualization modal view
The 'spinning' attribute is no longer directly manipulated, enhancing code maintainability. A deferred object method now manage the spinner status. The previous approach was causing the application to continuously re-render the visualization section. Everytime the spinner was changed it would force a re-render of the view causing duplicate graphs to be displayed.

* [ALS-6232] Update login buttons (#299)

* [ALS-6227] Contact Us opens in a new tab (#300)

* [ALS-6171] Show stigvars in open access but disable filtering (#298)

* [ALS-6171] Fix info modal disabling filter bug (#303)

* [ALS-6308] User feedback on Stigmatizing Variables (#304)

* [ALS-6308] link to variable-info-modal
* [ALS-6308] Add stigmatization status to variable info modal
* [ALS-6308] Updated stigmatized variable text in modal
* [ALS-6388] Pre-filter search results by stigmatizing & isOpenAccess
* [ALS-6308] Improved code readability in variable-info-modal-template
* [ALS-6385] Updating title text for search result filter icon

* Updated tooltip message for disabled filter variables (#309)

---------

Co-authored-by: GeorgeC <gcolon021@gmail.com>
Co-authored-by: Gcolon021 <34667267+Gcolon021@users.noreply.github.com>
Co-authored-by: James <Jamestp19@gmail.com>